HOCS-4341: add kubernetes deployment config (#3)
* HOCS-4340: add build drone pipeline

This change adds a initial build and push pipeline. This builds the java
 project, runs the tests and then creates and pushes the resultant
 image to Quay.

* HOCS-4341: add files directory

This change adds a file directory that is copied into the docker image
that will be used by the image to update. A default empty file has been
added but this should be replaced in-branch by a file to update.

* HOCS-4341: rework properties

The properties file was not currently processable by the spring service.
This change converts them into a readable yaml format for the service.

* HOCS-4341: add deployment drone pipeline

This change adds the required deployment steps into the drone script.
This contains separate pipeline steps for releasing to not-orod and prod
 environments.

* HOCS-4341: add job deployment config

This change adds the deployment config for the 'service'.

This job has a restartPolicy of never and also a backOffLimit of 1, to
reduce the chance of the image starting a second time.

If the job is completed the deploy script then deletes the job.

* HOCS-4341: update readme and props

This change updates the readme and changes the documentation on the
variables it requires. This also updates the documentation to tell the
user where to put the files.

An separate change to remove the SERVER_PORT as it's not a deployable
servlet has also been actioned.